refactor(Navbar): remove variant prop and determine user type dynamically and render on server
