+++ src/app/(landing_page)/layout.tsx
@@ -4,7 +4,7 @@ import Navbar from "@/components/Navbar";
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <AuthGuard guestOnly redirectPath="/main">
-      <Navbar variant="guest" />
+      <Navbar />
       {children}
       {/* <Footer /> */}
     </AuthGuard>
