+++ src/app/auth/layout.tsx
@@ -4,7 +4,7 @@ import AuthGuard from "@/components/AuthGuard";
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <AuthGuard guestOnly redirectPath="/main">
-      <Navbar variant="guest" />
+      <Navbar />
       {children}
     </AuthGuard>
   );
