+++ src/components/Navbar/Navbar.tsx
@@ -1,16 +1,8 @@
-"use client";
-
 import Link from "next/link";
-import { useState, useRef, useEffect } from "react";
 import Logo from "../Logo/Logo";
-import Avatar from "../Avatar/Avatar";
-import { RxHamburgerMenu } from "react-icons/rx";
-import { AuthService } from "@/services/AuthService";
-import { useRouter } from "next/navigation";
-
-interface Props {
-  variant: "tenant" | "landlord" | "analyst" | "guest";
-}
+import GuestRightSection from "./components/GuestRightSection";
+import UserRightSection from "./components/UserRightSection";
+import { getUserServerSide } from "../AuthGuard/AuthGuard";
 
 const navbarLinks = {
   tenant: [
@@ -38,7 +30,10 @@ const navbarLinks = {
   ]
 };
 
-export default function Navbar({ variant }: Props) {
+export default async function Navbar() {
+  const user = await getUserServerSide();
+  const variant = user?.userType.toLocaleLowerCase() || "guest";
+
   return (
     <nav className="bg-white shadow-sm">
       <div className="container mx-auto flex items-center justify-between py-3 px-6">
@@ -59,87 +54,12 @@ export default function Navbar({ variant }: Props) {
         </ul>
 
         {/* Right section */}
-        {variant === "guest" ? <GuestRightSection /> : <UserRightSection />}
+        {variant === "guest" ? (
+          <GuestRightSection />
+        ) : (
+          <UserRightSection pfpPath={user?.profilePictureUrl || ""} />
+        )}
       </div>
     </nav>
-  );
-}
-
-function GuestRightSection() {
-  return (
-    <div className="flex items-center space-x-4">
-      <Link
-        className="text-black-600 mx-2 sm:mx-4 capitalize tracking-wide hover:text-primary transition-all"
-        href="/auth/login"
-      >
-        Log in
-      </Link>
-      <Link
-        href={"/auth/register"}
-        className="font-medium tracking-wide py-2 px-4 border-2 border-primary text-primary bg-white outline-none rounded-l-full rounded-r-full capitalize hover:bg-primary hover:text-white transition-all "
-      >
-        Sign up
-      </Link>
-    </div>
-  );
-}
-
-function UserRightSection() {
-  const router = useRouter();
-  const [menuOpen, setMenuOpen] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-
-  // Close menu when clicking outside
-  useEffect(() => {
-    function handleClickOutside(event: MouseEvent) {
-      if (menuRef.current && !menuRef.current.contains(event.target as Node)) {
-        setMenuOpen(false);
-      }
-    }
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, []);
-
-  async function onLogout() {
-    await AuthService.signOut();
-    router.push("/");
-  }
-
-  return (
-    <div className="relative" ref={menuRef}>
-      <button
-        onClick={() => setMenuOpen((prev) => !prev)}
-        className="flex items-center space-x-4 border border-gray-300 rounded-full p-2 hover:bg-gray-50 cursor-pointer"
-      >
-        <RxHamburgerMenu className="size-7 p-1 mr-2 ml-1" />
-        <Avatar imageUrl="/pfp_placeholder.png" className="size-7" />
-      </button>
-
-      {/* Dropdown with transition */}
-      <div
-        className={`overflow-hidden absolute right-0 top-full mt-2 w-48 bg-white border border-gray-300 rounded-lg shadow-md z-50 transform transition-all duration-200 ease-out
-              ${menuOpen ? "opacity-100 scale-100 translate-y-0" : "opacity-0 scale-95 -translate-y-2 pointer-events-none"}
-            `}
-      >
-        <ul className="flex flex-col text-sm text-gray-700">
-          <li>
-            <Link
-              href="/edit-profile"
-              className="block px-4 py-2 hover:bg-gray-100"
-            >
-              Edit Profile
-            </Link>
-          </li>
-          <li>
-            <button
-              className="w-full text-left px-4 py-2 hover:bg-gray-100 cursor-pointer"
-              onClick={onLogout}
-            >
-              Logout
-            </button>
-          </li>
-        </ul>
-      </div>
-    </div>
   );
 }
